+++ tests/.lib/conf.sh
@@ -587,6 +587,7 @@ run_bg() {
 		return 1
 	fi
 
+	>&2 printf "PID: ${pid}\\n"
 	>&2 printf "${clr_ok}%s${clr_rst}\\n" "[OK]"
 	echo "${pid}"
 	print_test_datetime
@@ -660,7 +661,9 @@ wait_for_data_transferred() {
 				else
 					print_error "[Receive Error] Received data on ${recv_name} does not match expected data."
 				fi
-				kill_pid "${send_pid}" || true
+				if [ -n "${send_pid}" ]; then
+					kill_pid "${send_pid}" || true
+				fi
 				kill_pid "${recv_pid}" || true
 				print_test_datetime
 				exit 1
@@ -722,7 +725,9 @@ wait_for_data_transferred() {
 				else
 					print_error "[Receive Error] Received data on ${recv_name} does not match expected data."
 				fi
-				kill_pid "${send_pid}" || true
+				if [ -n "${send_pid}" ]; then
+					kill_pid "${send_pid}" || true
+				fi
 				kill_pid "${recv_pid}" || true
 				print_test_datetime
 				exit 1
@@ -771,7 +776,9 @@ wait_for_data_transferred() {
 				else
 					print_error "[Receive Error] Received data on ${recv_name} does not match expected data."
 				fi
-				kill_pid "${send_pid}" || true
+				if [ -n "${send_pid}" ]; then
+					kill_pid "${send_pid}" || true
+				fi
 				kill_pid "${recv_pid}" || true
 				print_test_datetime
 				exit 1
@@ -943,6 +950,7 @@ test_case_instance_is_stopped() {
 	fi
 	print_file "${name} STDERR" "${file_stderr}"
 	print_file "${name} STDOUT" "${file_stdout}"
+	run "ps -ef" || true
 	print_error "[${name} Error] ${name} is not stopped"
 
 	# cleanup