+++ tests/integration/10-mode---local_forward/000---tcp---www-forward.sh
@@ -153,19 +153,19 @@ run_test() {
 
 for curr_round in $(seq "${RUNS}"); do
 	#         server opts         client opts
-	run_test "--local ${RHOST}:${RPORT} ${THOST} ${TPORT} -vvvv" "-vvvv"  "1" "13" "${curr_round}" "${RUNS}"
-	#run_test "--local ${RHOST}:${RPORT} ${THOST} ${TPORT} -vvv " "-vvvv"  "2" "13" "${curr_round}" "${RUNS}"
-	#run_test "--local ${RHOST}:${RPORT} ${THOST} ${TPORT} -vv  " "-vvvv"  "3" "13" "${curr_round}" "${RUNS}"
-	#run_test "--local ${RHOST}:${RPORT} ${THOST} ${TPORT} -v   " "-vvvv"  "4" "13" "${curr_round}" "${RUNS}"
-	#run_test "--local ${RHOST}:${RPORT} ${THOST} ${TPORT}      " "-vvvv"  "5" "13" "${curr_round}" "${RUNS}"
+	run_test "--local ${RHOST}:${RPORT} --no-shutdown ${THOST} ${TPORT} -vvvv" "--no-shutdown -vvvv"  "1" "13" "${curr_round}" "${RUNS}"
+	#run_test "--local ${RHOST}:${RPORT} --no-shutdown ${THOST} ${TPORT} -vvv " "--no-shutdown -vvvv"  "2" "13" "${curr_round}" "${RUNS}"
+	#run_test "--local ${RHOST}:${RPORT} --no-shutdown ${THOST} ${TPORT} -vv  " "--no-shutdown -vvvv"  "3" "13" "${curr_round}" "${RUNS}"
+	#run_test "--local ${RHOST}:${RPORT} --no-shutdown ${THOST} ${TPORT} -v   " "--no-shutdown -vvvv"  "4" "13" "${curr_round}" "${RUNS}"
+	#run_test "--local ${RHOST}:${RPORT} --no-shutdown ${THOST} ${TPORT}      " "--no-shutdown -vvvv"  "5" "13" "${curr_round}" "${RUNS}"
 
-	#run_test "--local ${RHOST}:${RPORT} ${THOST} ${TPORT} -vvvv" "-vvv "  "6" "13" "${curr_round}" "${RUNS}"
-	#run_test "--local ${RHOST}:${RPORT} ${THOST} ${TPORT} -vvvv" "-vv  "  "7" "13" "${curr_round}" "${RUNS}"
-	#run_test "--local ${RHOST}:${RPORT} ${THOST} ${TPORT} -vvvv" "-v   "  "8" "13" "${curr_round}" "${RUNS}"
-	#run_test "--local ${RHOST}:${RPORT} ${THOST} ${TPORT} -vvvv" "     "  "9" "13" "${curr_round}" "${RUNS}"
+	#run_test "--local ${RHOST}:${RPORT} --no-shutdown ${THOST} ${TPORT} -vvvv" "--no-shutdown -vvv "  "6" "13" "${curr_round}" "${RUNS}"
+	#run_test "--local ${RHOST}:${RPORT} --no-shutdown ${THOST} ${TPORT} -vvvv" "--no-shutdown -vv  "  "7" "13" "${curr_round}" "${RUNS}"
+	#run_test "--local ${RHOST}:${RPORT} --no-shutdown ${THOST} ${TPORT} -vvvv" "--no-shutdown -v   "  "8" "13" "${curr_round}" "${RUNS}"
+	#run_test "--local ${RHOST}:${RPORT} --no-shutdown ${THOST} ${TPORT} -vvvv" "--no-shutdown      "  "9" "13" "${curr_round}" "${RUNS}"
 
-	#run_test "--local ${RHOST}:${RPORT} ${THOST} ${TPORT} -vvv " "-vvv " "10" "13" "${curr_round}" "${RUNS}"
-	#run_test "--local ${RHOST}:${RPORT} ${THOST} ${TPORT} -vv  " "-vv  " "11" "13" "${curr_round}" "${RUNS}"
-	#run_test "--local ${RHOST}:${RPORT} ${THOST} ${TPORT} -v   " "-v   " "12" "13" "${curr_round}" "${RUNS}"
-	#run_test "--local ${RHOST}:${RPORT} ${THOST} ${TPORT}      " "     " "13" "13" "${curr_round}" "${RUNS}"
+	#run_test "--local ${RHOST}:${RPORT} --no-shutdown ${THOST} ${TPORT} -vvv " "--no-shutdown -vvv " "10" "13" "${curr_round}" "${RUNS}"
+	#run_test "--local ${RHOST}:${RPORT} --no-shutdown ${THOST} ${TPORT} -vv  " "--no-shutdown -vv  " "11" "13" "${curr_round}" "${RUNS}"
+	#run_test "--local ${RHOST}:${RPORT} --no-shutdown ${THOST} ${TPORT} -v   " "--no-shutdown -v   " "12" "13" "${curr_round}" "${RUNS}"
+	#run_test "--local ${RHOST}:${RPORT} --no-shutdown ${THOST} ${TPORT}      " "--no-shutdown      " "13" "13" "${curr_round}" "${RUNS}"
 done