+++ tests/integration/20-options---nodns/100---tcp---server_nodns.sh
@@ -87,15 +87,15 @@ run_test() {
 # -------------------------------------------------------------------------------------------------
 
 for curr_round in $(seq "${RUNS}"); do
-	run_test "-l ${RHOST} ${RPORT} -n -vvvv     "  "1" "10" "${curr_round}" "${RUNS}"
-	#run_test "-l ${RHOST} ${RPORT} -n -vvv      "  "2" "10" "${curr_round}" "${RUNS}"
-	#run_test "-l ${RHOST} ${RPORT} -n -vv       "  "3" "10" "${curr_round}" "${RUNS}"
-	#run_test "-l ${RHOST} ${RPORT} -n -v        "  "4" "10" "${curr_round}" "${RUNS}"
-	#run_test "-l ${RHOST} ${RPORT} -n           "  "5" "10" "${curr_round}" "${RUNS}"
+	run_test "-l ${RHOST} ${RPORT} --no-shutdown -n -vvvv     "  "1" "10" "${curr_round}" "${RUNS}"
+	#run_test "-l ${RHOST} ${RPORT} --no-shutdown -n -vvv      "  "2" "10" "${curr_round}" "${RUNS}"
+	#run_test "-l ${RHOST} ${RPORT} --no-shutdown -n -vv       "  "3" "10" "${curr_round}" "${RUNS}"
+	#run_test "-l ${RHOST} ${RPORT} --no-shutdown -n -v        "  "4" "10" "${curr_round}" "${RUNS}"
+	#run_test "-l ${RHOST} ${RPORT} --no-shutdown -n           "  "5" "10" "${curr_round}" "${RUNS}"
 
-	#run_test "-l ${RHOST} ${RPORT} --nodns -vvvv"  "6" "10" "${curr_round}" "${RUNS}"
-	#run_test "-l ${RHOST} ${RPORT} --nodns -vvv "  "7" "10" "${curr_round}" "${RUNS}"
-	#run_test "-l ${RHOST} ${RPORT} --nodns -vv  "  "8" "10" "${curr_round}" "${RUNS}"
-	#run_test "-l ${RHOST} ${RPORT} --nodns -v   " " 9" "10" "${curr_round}" "${RUNS}"
-	#run_test "-l ${RHOST} ${RPORT} --nodns      " "10" "10" "${curr_round}" "${RUNS}"
+	#run_test "-l ${RHOST} ${RPORT} --no-shutdown --nodns -vvvv"  "6" "10" "${curr_round}" "${RUNS}"
+	#run_test "-l ${RHOST} ${RPORT} --no-shutdown --nodns -vvv "  "7" "10" "${curr_round}" "${RUNS}"
+	#run_test "-l ${RHOST} ${RPORT} --no-shutdown --nodns -vv  "  "8" "10" "${curr_round}" "${RUNS}"
+	#run_test "-l ${RHOST} ${RPORT} --no-shutdown --nodns -v   " " 9" "10" "${curr_round}" "${RUNS}"
+	#run_test "-l ${RHOST} ${RPORT} --no-shutdown --nodns      " "10" "10" "${curr_round}" "${RUNS}"
 done