+++ tests/smoke/200---tcp---keep_open/start_client_send.sh
@@ -1,9 +1,0 @@
-#!/bin/sh
-set -e
-set -u
-set -x
-
-RHOST="${1}"
-RPORT="${2}"
-
-printf "hi\\n" | "python${PYTHON_VERSION}" /usr/bin/pwncat -vvvvv "${RHOST}" "${RPORT}"