+++ tests/smoke/201---tcp---keep_open---kill_server---send_data/data/start_client_send.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+set -e
+set -u
+set -x
+
+RHOST="${1}"
+RPORT="${2}"
+
+printf "hi\\n" | "python${PYTHON_VERSION}" /usr/bin/pwncat --no-shutdown -vvvvv "${RHOST}" "${RPORT}"